refactor: move marketplace to Node app
