+++ imports/node-app/plugins/marketplace/i18n/fr.json
