+++ imports/node-app/plugins/payments-stripe/util/getStripeInstance.js
@@ -1,20 +1,22 @@
 import stripeNpm from "stripe";
-import packageJson from "/package.json";
 
 // This should not be customized per application.
 const APP_INFO = {
   name: "ReactionCommerce",
-  version: packageJson.version,
-  url: packageJson.url
+  url: "https://github.com/reactioncommerce/reaction"
 };
 
 /**
  * @name getStripeInstance
+ * @param {Object} context App context
  * @param {String} stripeApiKey Stripe API Key, see https://stripe.com/docs/keys
  * @returns {Object} The Stripe SDK object
  */
-export default function getStripeInstance(stripeApiKey) {
+export default function getStripeInstance(context, stripeApiKey) {
   const stripe = stripeNpm(stripeApiKey);
-  stripe.setAppInfo(APP_INFO);
+  stripe.setAppInfo({
+    ...APP_INFO,
+    version: context.appVersion
+  });
   return stripe;
 }
