+++ imports/node-app/plugins/payments-stripe/util/getStripeInstanceForShop.js
@@ -3,7 +3,6 @@ import Logger from "@reactioncommerce/logger";
 import getStripeInstance from "./getStripeInstance.js";
 import getStripePackageForShop from "./getStripePackageForShop.js";
 
-
 /**
  * @summary Given a shop ID, gets an instance of the Stripe API configured with that shop's API key.
  * @param {Object} context The context object, with `collections.Packages` on it
@@ -19,11 +18,11 @@ export default async function getStripeInstanceForShop(context, shopId) {
     const stripeAccessToken = (stripePkgSettings.connectAuth || {}).access_token;
     if (stripeAccessToken) {
       Logger.warn("Using a Stripe access_token instead of an api_key is deprecated. Please set an API Key.");
-      return getStripeInstance(stripeAccessToken);
+      return getStripeInstance(context, stripeAccessToken);
     }
 
     throw new ReactionError("not-configured", "Stripe is not configured properly. Please set an API Key.");
   }
 
-  return getStripeInstance(stripeApiKey);
+  return getStripeInstance(context, stripeApiKey);
 }
